Re-enable rapidcheck property tests
Fix the way we use CMake to find the rapidcheck library. Installing it
via `make install` is different than the previous Conan method and that
has led to the tests not compiling and not running.

Call Database::cleanup() explicitly in the db_adapter_property_test in
order to start with a fresh DB on every input.

Fix performance errors (object copies) reported by clang-tidy in
internal_node_property_tests.