Use big-endian and add fixed-sized lists in CMF
Change the byte order of integer serialization in CMF to big-endian.
Rationale for doing so is:
 * big-endian integers can be lexicographically ordered - that can be
   useful for having versioned CMF keys in an ordered DB, for example
 * less of reason is that network byte order is defined as big-endian

Add a static_cast to the unsigned version of the type being
deserialized when fetching the next byte. That fixes issues for 64 bit
types, because the left shift was promoted to an integer instead of the
correct 64 bit type.

Add support for the `fixedlist` type, allowing lists of arbitrary types,
but of a fixed size. No length is serialized before the actual list.
Maps to `std::array` in C++. In Python, size checks are done for lists,
not allowing lists of an incorrect type.

Remove redundant checks for buffer size in CMFDeserializer.map() and
.list() methods.

Add support for C++ CMF in CMake by slightly modifying the
`FindCMFC.cmake` file from Concord (credit to @robem). Add support for
namespace configuration.

Make the std::string overloads of (de)serialize `static inline`,
allowing the use of multiple CMF files in the same binary.

Add a cppgen test to verify integer serialization produces correct
results, in addition to round-trip serialization/deserialization. That
test ensures no systematic error is present both in serialization and
deserialization code. Make test_cppgen a first-class test and run it as
part of ctest.

Use negative and positive values for signed integers in test_cppgen.

Add sanitize g++ options when compiling cppgen tests in order to guard
against undefined behavior related to bitwise operations, memory
corruption and leaks:
 * -fsanitize=undefined
 * -fsanitize=address
 * -fsanitize=leak
 * -fno-sanitize-recover