Make test stable to include in CI. (#1061)
* Make test stable to include in CI.
Because the View Change timer is set to tick with a precision of 1/2
of the View Change Timeout, we usually get it to timeout later
than the specified period. In order to make the test pass consistently
we need to start the Replicas with a delay of at least 3/2 of the
View Change period.

* Fix bug with cancellation of the background constant load.
We cannot cancel the nursery in which we start the background
sender and the test itself, we need to only cancel the background
task itself instead.

* Stabilized test for random order of Replicas startup.
When we have a random order of Replica startup we might have a View Change
depending on startup time of Replica 0. Either with or without View Change
the Replicas have to agree on a View and stabilize (process Client requests).