Add option to direct_kv_db adapter which disables separate saving of KVs (#1066)
Besides the raw block itself direct_kv_db adapter saves also all KV
pairs from the block in the database. This feature doesn't play well
with S3 (which shares the adapter code), because there is a limitation
for the size of the keys written to the object store (around 250 for
minio and 1024 for AWS S3).
For this reason additional flag is added to the adapter constructor
which controls whether to save the KV pairs AND the raw block or only
the raw block.