Initial implementation of a Byzantine replica
The idea of this change is to propose an initial implementation of a Byzantine replica, or in
other words, an extension of a regular replica where it's possible to inject different
Byzantine behavior by extending the default logic.

What this change proposes tries to achieve is basically:
- Create an extension of bftengine::ReplicaImp to expose some interesting methods that can be
overridden in order to inject some Byzantine behavior.
- Create an extension of kvbc::ReplicaImp to override some methods in order to instantiate the
new ByzantineBftEngineReplicaImp.
- Add an initial test case that makes use of the Byzantine classes.

See the README file and the comments inside the ByzantineReplicaImp.cpp for some more details.

Small disclaimer: I'm not sure whether this set of changes could or should be merged "as it is"
or not, probably not :) But with the proper feedback, addressing the comments and perhaps
splitting the changes in different reviews if needed, it could be a good enough starting point.