BCStateTran: Add ST FetchBlockMsg <-> ItemDataMsg  latency histogram
Here we add s FetchBlockMsg <->ItemDataMsg  latency histogram.
This message exchange process is between destination (collecting)
replica and a chosen source. Every time such a process ends, a
cumulative percentile latency histogram is snapshotted and logged.

To record the time passed between send and reply, we use an
AsyncTimeRecorderMap atomic recorder. When FetchBlockMsg is sent, we
call start() and insert the message sequence number. When a valid
ItemDataMsg reply arrives, we call end() with the reply sequence number.