Storage Categorization: Add BlockMerkle category
The block merkle category operates by only storing a single key in the
sparse merkle tree per block. The key is the block id, and the value is
the hash of the concatenation of the hashes of all the keys and values
(including deleted keys), that were updated in the merkle tree for the
given block.

Direct lookup is enabled by versioned key lookup on hashed values.
Values can be looked up directly if the block id is known for the value,
or the latest value can be retrieved. MultiGet is also supported. In
order to avoid the computational overhead of RocksDB seeks, point
lookups of the latest value are done by maintaining an index for the
latst block id of each key in a separate column family.

Unit tests were added that mainly verify the behavior required by direct
gets and puts.

A follow up commit will add support for deletes and pruning behavior.
Unit tests will be added for this behavior, and to cover the block
merkle nodes and hashes appropriately.

Lastly, category type suffixes were changed from *UpdatesData* and
*UpdatesInfo* to *Input* and *Output*, respectively.