Implement common category read iface for Immutable
Implement the common category read interface for the
ImmutableKeyValueCategory. That includes:
 * get the value of a key at a specific version
 * get the value of a key at the latest version
 * get multiple keys at specific versions
 * get multiple keys at their latest versions
 * get the latest version of a key
 * get the latest versions of multiple keys

Introduce value types for the different categories.

Add an ability to partially deserialize immutable values by expoiting
CMF's canonical format. Use the `ImmutableDbVersion` CMF message that
contains the version only and deserialize it from a `ImmutableDbValue`
buffer that contains both version and value. That way we don't have to
copy the value when we only care about the version.