Check the complaints placed inside the View Change message for higher View than the Replica's current. (#1119)
This handles the case where we have a Replica behind more than 1 View from the others.
This is an improvement, the behind Replica can catch-up by the F+1 View Change messages that it will
receive from peers. We can though move to the agreed view earlier based on the F+1 signed Complaints.