Support TaggedVersion in VersionedKeyValueCategory
Unify categories by supporting TaggedVersion in
VersionedKeyValueCategory.

Optimize tombstone values by not using a variant type. Instead, use a
simple flag to designate a value is a tombstone. The value of a
tombstone is the empty buffer. Rationale is that the variant adds an
overhead of 4 bytes for its type.

Handle deleted keys in deleteLastReachableBlock() by preserving the
deleted flag when seeking back to a previously deleted key version.

Optimize multiGetLatest() by not including deleted (or not found) keys
in the values column family lookup. Implementation is borrowed from the
BlockMerkleCategory.

Add unit tests to verify the new behavior.