Add deleteGenesisBlock for BlockMerkle category
We delete keys when we prune a block and the version of the key is not
the latest version. If the key is the latest version we call it the
`active` key.

When all keys in a block are inactive we remove the block from the
merkle tree. Otherwise we write a new version of the block in the merkle
tree, as well as metadata indexes that contain the active keys by block
id (`PrunedBlock` msg), and the pruned block version by hashed key. This
metadata is necessary for proving the still active keys as well as
allowing garbage collection of prior active keys and
block-rewriting/deletion in the merkle tree to be postponed until prune
operations. Essentially, we make `add` faster at the cost of
`deleteGenesisBlock` and some small storage overhead. This work is
performed in the `writePrunedBlock` and `rewriteAlreadyPrunedBlocks`
methods.

When we update the merkle tree, we now save any stale keys so we can
delete them when they are no longer referenced. This is necessary for
the safety of the sparse copy-on-write merkle tree. Nodes in un-pruned
tree versions must be able to retrieve data from nodes that may have
been generated in blocks that were already pruned. We keep these nodes
available until the last referencing tree version gets pruned.

During pruning, new tree versions are created and nodes are marked
stale. These nodes are created because the merkle tree is modified by
either inserting an updated version of a block id and hash for a pruned
block with active keys, or the pruned block has no active keys, in which
case it is removed from the tree. These tree versions are not tied to
any block, since we don't create new blocks when we prune a block.
However, the stale nodes that get generated as a result of deleting
blocks must be garbage collected at some point. After pruning, when a
new block is added, it will be at a version greater than those stale
nodes. When this new block is eventually pruned, all prior stale
versions are finally deleted from the database. This work is performed
by the `deleteStaleData` method and tested in `prune_many_nodes`.

In order to make the deletion of stale nodes as a result of pruning many
blocks in a row before adding a new block efficient, we do it in atomic
batches in the `deleteStaleBatch` method. These batches are deleted
directly and not returned as one giant batch to the caller because:
 1. The stale keys do not need to all be deleted in one atomic batch.
 2. There can be upwards of a million versions to delete for large,
 uninterrupted pruning sessions done during maintenance windows.

A serious safety issue was discovered in testing, due to the use of
thread local buffer references in `detail::serialize`. When a reference
to the buffer is used in a `batch.put` operation for both a key and a
value of the same type, the value overwrites the key. This also occurs
in any put operations with multiple key references. This is extremely
dangerous and can result in different keys being written if a temporary
without a reference is used in a serialized key vesion. For example, the
difference between `auto x = serialize(key)` and
`batch.put(serialize(key), serialize(val))` is significant. Therefore
the `serialize` wrapper was changed to always allocate a new buffer. The
old behavior is preserved when users opt into it knowingly with
`serialiizeThreadLocal`.