Integrate the categorized KVBC interfaces
Implement the categorized KVBC interfaces in the KVBC replica and
integrate them into the whole codebase. Deprecate the non-categorized
interfaces.

Define a Concord and Concord-BFT internal category of type
VersionedKeyValueCategory. It can be used for various kinds of metadata.

Serialize the last sequence number in BlockMetadata as an 8 byte
big-endian integer. Make sure it is used accordingly in Concord.

Add USE_S3_OBJECT_STORE as a CMake option. Make sure the read-only
replica tests are not run when this option is false.

Use the categorized KVBC for normal read-write ReplicaImps. Use the
IDbAdapter instance for read-only ones.

Make sure the v1DirectKeyValue::DBAdapter::saveKvPairsSeparately_ flag
is honored and blocks with the new (and incompatible) categorization
format are not deserialized during addRawBlock().

Make IReplica::addBlockToIdleReplica() compatible with the categorized
KVBC. Also, throw an exception in case it fails instead of returning a
Status.

TODOs in subsequent commits:
 * Use getSlice() in the categorized blockchain/kv_blockchain code where
   possible (for performance)
 * Either remove or handle the USE_ROCKSDB compile flag correctly
 * Simplify Updates::categoryUpdates()
 * Support both VersionedKeyValue and BlockMerkle categories in
   TesterReplica
 * Consider running with different category types (or a mix) in Apollo
   tests
 * Consider making IReader::getBlockUpdates() return an optional
 * Consider an assert in ReplicaImp::putBlock() when an existing and
   different block is found in storage during ST
 * If possible, change the way we test for linearizability in Apollo and
   get rid of the linear InternalCommandsHandler::getAtMost()
 * Consider implementing the persistence-mode TesterReplica option via
   an in-memory RocksDB instance