TLS: Remove unnecessary lock contention for writes
Prior to this change, every send required acquiring two locks: One for
the actual per-conneciton queue being written into, and one global lock
across all connections. This latter lock was required because each queue
was created inside each AsyncTlsConnection and therefore the lifetime of
the queue was directly tied to the lifetime of the connection. Since
AsyncTlsConnections are dynamically created and destroyed, and only
become valid upon successful TLS authentication, any attempts to push onto the
associated queue must first acquire the connections lock to get access
to the queue if one exists. The connection lock therefore essentially
serializes all writes, even if they are for separate sockets.

However, since we know there are a specific number of fixed destinations
provided by config, we can create a static map of queues by destination
and decouple their lifetimes from the dynamically created connections
themselves. When a connection is established we associate the proper queue
with the connection. This allows us only to take a lock per queue, and
only when pushing and popping from the queue. Thus, we get greater write
concurrency, and also shorter lock times.

In order to prevent unnecessary memcopy calls in `sendAsyncMessage`,
when there is no valid connectioon, we maintain an atomic per queue
which says if there is a connection. This atomic is optimistic, and its
possible that the connection is destroyed after the check. But this is
ok, as it just means in that rare case we do an extra copy and push onto
the queue.