Use new histogram registration API for preprocessor
Also, remove the try/catch wrapper around registration of the component.
This will not work, as any defined recorders will be separate instances
and will not log to the existing registered recorders. In essence that
data will be silently dropped. The registration code was actually
changed to prevent this misuse by terminating in case of duplicates. If
you have multiple instances they should be registered independently or
there should be a central recorder that is a shared variable. An atomic
recorder can be used in this case.

* Allow clearing of performance handler registration during tests.
* Clear diagnostics handlers in perf tests.