Diagnostics: Terminate when registering dup handlers
Registration of handlers occurs using a singleton and comes from
multiple threads. We can't be sure the context that we are running in
and a joinable thread may not properly catch the thrown exception and
cause the process to exit with "terminate called without an active
exception".

Since a duplicate registration is a hard failure, just log it explicitly
to make diagnostics easier and terminate immediately.