Start of block merkle docs
I used draw.io to create these drawings and recommend using it for our
opensource documentaiton. The justification is that it's a web based
platform that is free to use and cross platform. You can save the images
as svg or png with the instructions to create them inside. Then draw.io
can reload them and they are editable. This enables open source
collaboration on our docs.

An overview doc describing things at a logical level exists, and then a
doc containing more implementation details with links is provided.