Remove pending requests after the sequence number commit (instead of execution)
The execution is a detemenistic part in the protocol that is not depended on other replicas data. Thus it make sense not to take the execution time in account for complaining the primary.
In this PR we remove the pending request after the commit instead of after the execution.
We expect this change to stablize VC and prevent unnecessary VCs