Remove TesterReplica's "-p" option
No longer support the "-p" option in TesterReplica as the kvbc library
is dependent on RocksDB. If in-memory operation is needed, TesterReplica
can be run with a path that points to a `tmpfs` or `ramfs` disk.

In addition, remove usage of the "-p" option in Apollo. Rationale is
that non-persistent mode is not a real use case and, therefore, is not
worth testing.