Handle non-atomic RocksDB column family creation (#1149)
RocksDB creates column families in 2 non-atomic steps:
 1. Create the column family in the DB.
 2. Create the column family in the options file by persisting the
    options provided by the user.

Reference:
https://github.com/facebook/rocksdb/blob/v6.8.1/db/db_impl/db_impl.cc#L2187

If the process is stopped between these steps, there could be a mismatch
between them. That could lead to:
 1. The DB open process failing due to missing column families in the
    call to Open() as the cf_descs variable that contains column family
    names and options is loaded from the options file.
 2. Losing the column family options the user provided.

In order to solve it, track incompletely created column families (i.e.
ones that have mismatching information about them in the DB and the
options file). If such a family is empty (i.e. has no keys inside), drop
it so that user code can re-create it. If not empty, throw an exception.
Rationale for throwing is:
 * we want to see if that happens during testing and by just logging we
   might miss it
 * we don't want to run the system with possibly incorrect column family
   options as that can lead to slowdowns, affecting the whole cluster

The test causing the issue to show up is `test_skvbc_persistence.py`
when used with key categorization.