Implement deleteLastReachable for the BlockMerkleCategory
This is required for sync between concord-bft metadata and the
application data in the `BlockMerkleCategory` on startup after a crash.

All tree nodes and stale indexes are removed for the block being
removed. Additionally all keys and version indexes are removed. The code
to do this latter part is based on code for the
`VersionedKeyValueCategory` by @dartdart26.

One important thing to keep in mind is around writing `DbValue` CMF
data. We write it as a `DbValueHeader` and the raw value passed in by
the caller as two separate slices, rather than forcing an extra copy
during serialization. We can do this using the RocksDB API enhancements
added in #1166. When we read out the value it comes back as one
serialized `DbValue`.

I also added fixes for the review comments made in #1143 and cleaned up
some comments.