Add multi-send method to TLS layer to avoid excess copies in bft client
This patch modifies ICommunication interface for two main reasons:
- Avoid unnecessary memory copying when the bft client sends messages
to all replicas. In the current implementation the message is copied for
each replica instance.
- Remove unnecessary use of raw pointers in the code.

How this is achieved:

Add new method to ICommunication - multiSendMessage(). It receives a raw
message buffer in form of std::vector<uint8_t> and takes ownership of
it. Internally the vector is converted to shared_ptr and it is shared
between all TLS connections. This approach avoids copying the same
message multiple times.

sendAsyncMessage() from ICommunication also accepts rvalue ref to
std::vector<uint8_t>. For consistency it is internally converted to
shared_ptr.

These changes lead to some additional modifications:
1. struct OutgoingMsg accepts vector instead of raw pointer. A copy of
the vector is made, because each message consists of LEN + PAYLOAD. To
construct the correct message structure a new chunk of memory needs to
be allocated.

2. The implementation of TlsTCPCommunication is changed. Now OutgoingMsg
is created from the raw buffer passed by the upper layer. This OutgoingMsg
is passed to the implementation. The benefit is that memory for the
message is allocated only once.

3. WriteQueue works with shared_ptr<OutgoingMsg> instead of OutgoingMsg
directly. Additionally OutgoingMsg is not created in the WriteQueue
anymore. The message is created in TlsTcpImpl send functions.

4. Add implementations for sendAsyncMessage and multiSendMessage to
PlainUDPCommunication and FakeCommunication.