Optimize BlockMerkleCategory::deleteGenesisBlock
Before this change we sometimes created two new merkle tree versions for
a single deleted block. With this change we now always only create 1.

Also, enhance the tests for `deleteGenesisBlock`. We explicitly check
that the stale leaf nodes and actual leaf nodes are what we expect after
each call to `add` and `deleteGenesisBlock`.

Lastly, add `deleteGenesisBlock` to the test fixture to reduce
boilerplate.