Beginnings of KVBC storage benchmarking tool
This tool is intended to allow benchmarking and tuning the new
`kvbc_blockchain` code. We generate a bunch of random data up front, and
then add blocks containing that data. To simulate pre-execution we also
perform conflict detection before each block addition. Right now
conflict detection is performed via a single multiget, since it is
possible to batch all conflict reads by taking the union of the keys in
each transaction in a batch. This is easily changed as needed.

This preliminary implementation does not use the actual
`kvbc_blockchain` API yet. It instead puts all keys into the
`BlockMerkle` category. The implementation was done this way for ease
of getting the scaffolding for the tool in place and because the full
`kvbc_blockchain` code is not merged in yet on master.

The code currently:
 * Generates random `BlockMerkleInput` up to a configurable memory limit
 * Generates a configurable set of read keys for use in pre-execution
 and conflict detection
 * Performs simulated conflict detection checks
 * Provides tunables for number of blocks, batch size, etc...
 * Adds a histogram for the time taken to add a block, and one for
 conflict detection
 * Runs a diagnostic server so stats can be retreived with concord-ctl
   during benchmark runs
 * Dumps the rocksdb counters at the end of a run

There's still more to do, including:
 * Using the real `kvbc_blockchain` interface
 * Adding support for concurrent reads for the duration of block adds to
   simulate pre-execution
 * Allowing disabling of pre-execution simulation, and inlining reads
   into post-execution.

This commit also includes a significant change to the RocksDB
`NativeClient`, such that configuration of the database can be performed
largely via a rocksdb configuration file. This is a change to the
`UserOptions` mechanism, which now takes a filepath for a RocksDB
configuration file, as well as a callback to run to apply any
configuration that cannnot be performed by a RocksDB initialization
file. This latter configuration is applied **after** the configuration
file is loaded. A helper function, `openRocksDB`, was added to the
`IDBClient` based `rocksdb_client` where shared code is possible among
file based and non-file based configuration.

`kvbcbench` runs using the included configuration file, and further
configuration is applied via the `completeRocksdbConfiguration` callback
in `main.cpp`. A 4GB LRU block cache is shared among all column families
for simplicity, and as a starting point for tuning. Additionally, each
column family has a bloom filter, and the `block_merkle_key_versions`
column family is optimized for point lookup, since we will never do
range scans on it. All of this configuration is only possible in code,
and so it is performed in the callback.

Only column families for the `BlockMerkleCategory` are
included in the configuration file since `kvbcbench` only uses that
category currently. Column families are created if they do not already
exist. As the `block_merkle_keys` column family contains the bulk of the
data, it is tuned to have a 512MB
write buffer, which creates an 8GB level 0 and level
1(`max_bytes_for_level_base`) based on the
multipliers: `level0_file_num_compaction_trigger` and
`min_write_buffer_number_to_merge`. Additionally, we limit files to 10 in
level 0 and level 1 with a `target_file_size_base` 1/10 the size
of `max_bytes_for_level_base`. We also cache index and filter blocks and
pin them in memory. We do this for the key versions as well, as we want
reads optimized.

It should be noted, that this configuration is *only* used by the
benchmarking tool, and is a starting point for further experiments.

More information about why these settings were changed can be found in the [RocksDB
Tuning
Guide](https://github.com/facebook/rocksdb/wiki/RocksDB-Tuning-Guide).